indexer/requirements.txt
alembic==1.13.0
fastapi==0.104.1
gunicorn==21.2.0
uvicorn==0.24.0.post1
loguru==0.7.2
pydantic==2.5.2
pydantic-settings==2.1.0
tvm-valuetypes==0.0.10
psycopg2-binary==2.9.9
SQLAlchemy==1.4.46
asyncpg==0.29.0
SQLAlchemy-Utils==0.41.1
requests==2.31.0
pytonlib==0.0.51
uvloop==0.19.0
httpx==0.25.2
fastapi-cache2[redis]==0.2.1
